pyproject.toml
[project]
name = "arc-gemm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-tvm-ffi>=0.1.10",
    "ninja>=1.13.0",
    "numpy>=2.4.4",
    "torch>=2.11.0",
    "typer>=0.25.0",
]

[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["arc*"]
